add themes to local storage

+++ app.js
@@ -1,27 +1,28 @@
-const tasks = [{
-        _id: '5d2ca9e2e03d40b326596aa7',
-        completed: true,
-        body: 'Occaecat non ea quis occaecat ad culpa amet deserunt incididunt elit fugiat pariatur. Exercitation commodo culpa in veniam proident laboris in. Excepteur cupidatat eiusmod dolor consectetur exercitation nulla aliqua veniam fugiat irure mollit. Eu dolor dolor excepteur pariatur aute do do ut pariatur consequat reprehenderit deserunt.\r\n',
-        title: 'Eu ea incididunt sunt consectetur fugiat non.',
-    },
-    {
-        _id: '5d2ca9e29c8a94095c1288e0',
-        completed: false,
-        body: 'Aliquip cupidatat ex adipisicing veniam do tempor. Lorem nulla adipisicing et esse cupidatat qui deserunt in fugiat duis est qui. Est adipisicing ipsum qui cupidatat exercitation. Cupidatat aliqua deserunt id deserunt excepteur nostrud culpa eu voluptate excepteur. Cillum officia proident anim aliquip. Dolore veniam qui reprehenderit voluptate non id anim.\r\n',
-        title: 'Deserunt laborum id consectetur pariatur veniam occaecat occaecat tempor voluptate pariatur nulla reprehenderit ipsum.',
-    },
-    {
-        _id: '5d2ca9e2e03d40b3232496aa7',
-        completed: true,
-        body: 'Occaecat non ea quis occaecat ad culpa amet deserunt incididunt elit fugiat pariatur. Exercitation commodo culpa in veniam proident laboris in. Excepteur cupidatat eiusmod dolor consectetur exercitation nulla aliqua veniam fugiat irure mollit. Eu dolor dolor excepteur pariatur aute do do ut pariatur consequat reprehenderit deserunt.\r\n',
-        title: 'Eu ea incididunt sunt consectetur fugiat non.',
-    },
-    {
-        _id: '5d2ca9e29c8a94095564788e0',
-        completed: false,
-        body: 'Aliquip cupidatat ex adipisicing veniam do tempor. Lorem nulla adipisicing et esse cupidatat qui deserunt in fugiat duis est qui. Est adipisicing ipsum qui cupidatat exercitation. Cupidatat aliqua deserunt id deserunt excepteur nostrud culpa eu voluptate excepteur. Cillum officia proident anim aliquip. Dolore veniam qui reprehenderit voluptate non id anim.\r\n',
-        title: 'Deserunt laborum id consectetur pariatur veniam occaecat occaecat tempor voluptate pariatur nulla reprehenderit ipsum.',
-    },
+const tasks = [
+    // {
+    //     _id: '5d2ca9e2e03d40b326596aa7',
+    //     completed: true,
+    //     body: 'Occaecat non ea quis occaecat ad culpa amet deserunt incididunt elit fugiat pariatur. Exercitation commodo culpa in veniam proident laboris in. Excepteur cupidatat eiusmod dolor consectetur exercitation nulla aliqua veniam fugiat irure mollit. Eu dolor dolor excepteur pariatur aute do do ut pariatur consequat reprehenderit deserunt.\r\n',
+    //     title: 'Eu ea incididunt sunt consectetur fugiat non.',
+    // },
+    // {
+    //     _id: '5d2ca9e29c8a94095c1288e0',
+    //     completed: false,
+    //     body: 'Aliquip cupidatat ex adipisicing veniam do tempor. Lorem nulla adipisicing et esse cupidatat qui deserunt in fugiat duis est qui. Est adipisicing ipsum qui cupidatat exercitation. Cupidatat aliqua deserunt id deserunt excepteur nostrud culpa eu voluptate excepteur. Cillum officia proident anim aliquip. Dolore veniam qui reprehenderit voluptate non id anim.\r\n',
+    //     title: 'Deserunt laborum id consectetur pariatur veniam occaecat occaecat tempor voluptate pariatur nulla reprehenderit ipsum.',
+    // },
+    // {
+    //     _id: '5d2ca9e2e03d40b3232496aa7',
+    //     completed: true,
+    //     body: 'Occaecat non ea quis occaecat ad culpa amet deserunt incididunt elit fugiat pariatur. Exercitation commodo culpa in veniam proident laboris in. Excepteur cupidatat eiusmod dolor consectetur exercitation nulla aliqua veniam fugiat irure mollit. Eu dolor dolor excepteur pariatur aute do do ut pariatur consequat reprehenderit deserunt.\r\n',
+    //     title: 'Eu ea incididunt sunt consectetur fugiat non.',
+    // },
+    // {
+    //     _id: '5d2ca9e29c8a94095564788e0',
+    //     completed: false,
+    //     body: 'Aliquip cupidatat ex adipisicing veniam do tempor. Lorem nulla adipisicing et esse cupidatat qui deserunt in fugiat duis est qui. Est adipisicing ipsum qui cupidatat exercitation. Cupidatat aliqua deserunt id deserunt excepteur nostrud culpa eu voluptate excepteur. Cillum officia proident anim aliquip. Dolore veniam qui reprehenderit voluptate non id anim.\r\n',
+    //     title: 'Deserunt laborum id consectetur pariatur veniam occaecat occaecat tempor voluptate pariatur nulla reprehenderit ipsum.',
+    // },
 ];
 
 (function(arrOftasks) {
@@ -30,60 +31,16 @@ const tasks = [{
         return acc;
     }, {});
     const themes = {
-        default: {
-            '--base-text-color': '#212529',
-            '--header-bg': '#007bff',
-            '--header-text-color': '#fff',
-            '--default-btn-bg': '#007bff',
-            '--default-btn-text-color': '#fff',
-            '--default-btn-hover-bg': '#0069d9',
-            '--default-btn-border-color': '#0069d9',
-            '--danger-btn-bg': '#dc3545',
-            '--danger-btn-text-color': '#fff',
-            '--danger-btn-hover-bg': '#bd2130',
-            '--danger-btn-border-color': '#dc3545',
-            '--input-border-color': '#ced4da',
-            '--input-bg-color': '#fff',
-            '--input-text-color': '#495057',
-            '--input-focus-bg-color': '#fff',
-            '--input-focus-text-color': '#495057',
-            '--input-focus-border-color': '#80bdff',
-            '--input-focus-box-shadow': '0 0 0 0.2rem rgba(0, 123, 255, 0.25)',
-        },
-        dark: {
-            '--base-text-color': '#212529',
-            '--header-bg': '#343a40',
-            '--header-text-color': '#fff',
-            '--default-btn-bg': '#58616b',
-            '--default-btn-text-color': '#fff',
-            '--default-btn-hover-bg': '#292d31',
-            '--default-btn-border-color': '#343a40',
-            '--default-btn-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
-            '--danger-btn-bg': '#b52d3a',
-            '--danger-btn-text-color': '#fff',
-            '--danger-btn-hover-bg': '#88222c',
-            '--danger-btn-border-color': '#88222c',
-            '--input-border-color': '#ced4da',
-            '--input-bg-color': '#fff',
-            '--input-text-color': '#495057',
-            '--input-focus-bg-color': '#fff',
-            '--input-focus-text-color': '#495057',
-            '--input-focus-border-color': '#78818a',
-            '--input-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
-            '--li-background-color': '#5c5d5f',
-            // --text-color: white;
-        },
         light: {
-            '--base-text-color': '#212529',
             '--header-bg': '#fff',
             '--header-text-color': '#212529',
-            '--default-btn-bg': '#fff',
-            '--default-btn-text-color': '#212529',
-            '--default-btn-hover-bg': '#e8e7e7',
-            '--default-btn-border-color': '#343a40',
+            '--default-btn-bg': 'rgb(172, 172, 172)',
+            '--default-btn-text-color': '#fff',
+            '--default-btn-hover-bg': 'rgb(196, 196, 196)',
+            '--default-btn-border-color': 'rgb(172, 172, 172)',
             '--default-btn-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
-            '--danger-btn-bg': '#f1b5bb',
-            '--danger-btn-text-color': '#212529',
+            '--danger-btn-bg': '#dc3545',
+            '--danger-btn-text-color': '#fff',
             '--danger-btn-hover-bg': '#ef808a',
             '--danger-btn-border-color': '#e2818a',
             '--input-border-color': '#ced4da',
@@ -93,9 +50,42 @@ const tasks = [{
             '--input-focus-text-color': '#495057',
             '--input-focus-border-color': '#78818a',
             '--input-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
+            '--li-background-color': 'white',
+            '--text-color': 'black',
+            '--body-color': 'rgb(228, 224, 224)',
+            '--card-bg-color': 'white'
+
+
         },
+        dark: {
+            '--base-text-color': '#212529',
+            '--header-bg': '#343a40',
+            '--header-text-color': '#fff',
+            '--default-btn-bg': '#343a40',
+            '--default-btn-text-color': '#fff',
+            '--default-btn-hover-bg': '#292d31',
+            '--default-btn-border-color': '#343a40',
+            '--default-btn-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
+            '--danger-btn-bg': '#b52d3a',
+            '--danger-btn-text-color': '#fff',
+            '--danger-btn-hover-bg': '#88222c',
+            '--danger-btn-border-color': '#88222c',
+            '--input-border-color': '#ced4da',
+            '--input-bg-color': '#b3b7bb',
+            '--input-text-color': '#495057',
+            '--input-focus-bg-color': '#e0e4e7',
+            '--input-focus-text-color': '#495057',
+            '--input-focus-border-color': '#78818a',
+            '--input-focus-box-shadow': '0 0 0 0.2rem rgba(141, 143, 146, 0.25)',
+            '--li-background-color': '#343a40',
+            '--text-color': 'white',
+            '--body-color': '#495057',
+            '--card-bg-color': '#4e545a'
+        },
+
     };
-    let lastSelectedTheme = 'default';
+    let lastSelectedTheme = localStorage.getItem('app_theme') || 'default';
+    console.log(lastSelectedTheme)
 
     //Elements UI
     const listContainer = document.querySelector('.tasks-list-section .list-group')
@@ -106,6 +96,7 @@ const tasks = [{
 
 
     //Events
+    setTheme(lastSelectedTheme)
     renderAllTasks(objOfTasks);
     form.addEventListener('submit', onFormSubmitHandler)
     listContainer.addEventListener('click', onDeleteHandler)
@@ -208,23 +199,26 @@ const tasks = [{
     }
 
     function onDeleteHandler({ target }) {
-        const parent = target.closest('[data-task-id]')
-        const id = parent.dataset.taskId;
-        const confirmed = deleteTask(id);
-        deleteTaskFromHtml(parent, confirmed)
-
+        if (target.classList.contains('delete-btn')) {
+            const parent = target.closest('[data-task-id]')
+            const id = parent.dataset.taskId;
+            const confirmed = deleteTask(id);
+            deleteTaskFromHtml(parent, confirmed)
+        }
 
     }
 
     function onThemeSelectHandler(e) {
         const selectedTheme = themeSelect.value;
-        const isConfirmed = confirm(`Вы действетельно хотите изменить тему: ${selectedTheme}`);
-        if (!isConfirmed) return;
         setTheme(selectedTheme);
+        lastSelectedTheme = selectedTheme;
+        localStorage.setItem('app_theme', selectedTheme);
     }
 
     function setTheme(name) {
         const selectedThemeObj = themes[name];
+
+        console.log(selectedThemeObj)
 
         Object.entries(selectedThemeObj).forEach(([key, value]) => {
             console.log(key, value)
